id_sd_sdl: Make QueueAudio minimum buffer size configurable
The SDL_QueueAudio backend supports setting a "minimum buffer size": if
the amount of queued audio drops too low, more audio will be rendered to
bring it back up.

In general, this should never be necessary, and can cause timing
problems, so it's 0 by default. But if you're getting lots of crackling,
and don't mind the occasional note being held for way too long, feel
free to increase this to any number <= 512.

This can be done by adding this to OMNISPK.CFG:
sd_sdl_queueAudioMinBufSize = 256