id_sd_sdl: Make register write delays configurable
Currently, we delay a few samples after every register write. This can
be confusing, particularly with Nuked_OPL3, which also delays a couple
of samples when OPL3_WriteRegBuffered() is used. It's also ugly to have
to track all of the extra samples queued here and there.

Make the number of delayed samples configurable with:
sd_oplWriteDelay = <num>

It's still based of the sample rate by default, otherwise put the number
of samples you want, or 0 to disable it entirely.

You can also disable/enable the OPL3_WriteRegBuffered() function with
sd_nukedBufferWrites. By default, it's only enabled when
sd_oplWriteDelay is nonzero.

A PR to make NukedOPL3's buffering configurable exists here:
https://github.com/nukeykt/Nuked-OPL3/pull/27

Note that NukedOPL3's buffering is based in the 49716Hz samples, not the
resampled values, like our sd_oplWriteDelay.