id_sd_sdl: Add config option to override driver
In addition to the SDL_AUDIODRIVER environment variable, allow setting
the audio driver from OMNISPK.CFG to be consistent with other config
options.

We're overriding the default on Win32 for OSI, due to WASAPI being
somewhat dodgy, but we won't for the normal Omnispeak builds for now, as
they support SDL1 as well.

Just set the following in OMNISPK.CFG:
sd_sdl_audioDriver = "directsound"