ck_main: Support HighDPI on Win32
Windows doesn't support per-window DPI awareness, so use the new SDL
hint to force all windows to be DPI aware.

If this is broken, it can be overriden on the command line with
set SDL_WINDOWS_DPI_SCALING=0

And High-DPI support can be made to work by forcing it in the Windows
compatibility settings anyway.