sd_opl2alsa: Don't include x86-specific headers
As part of an old direct IO port access feature which I gave up on long
ago, id_sd_opl2alsa.c included <sys/io.h>.

This header only exists on x86 platforms, so Omnispeak wouldn't build
on, e.g., aarch64 if WITH_ALSA was declared. It's pretty unlikely that
a non-x86 machine has an OPL3 compatible soundcard, but that's no reason
not to support it if it does.

Of course, all of this should go away and be replaced with liboplhw, but
for now, let's fix it.

Thanks to Martin Hauke for creating openSUSE packages, whose failed
aarch64 build uncovered this issue:
https://build.opensuse.org/package/live_build_log/home:mnhauke:games/omnispeak/openSUSE_Tumbleweed/aarch64