id_in: Add 'text input' mode for, e.g., on-screen keyboards
This makes text-entry screens (savegames, F10+W, etc) pop up an
on-screen keyboard on some systems. The big one here is the Steam Deck:
modern-enough versions of SDL2 (including the one in the Steam Linux
Runtime) will detect the Steam Deck, and enable its onscreen keyboard.

Some Wayland backends might do this, too.

Note that the "reason" and "existing" arguments are unused: they're
there in case we want to provide additional detail to the input method,
such as some pre-existing text to edit, or a "name" for the field being
edited. The old Steam controller "Gamepad" interface supported this, but
the new floating one doesn't seem to:
https://partner.steamgames.com/doc/api/ISteamUtils#ShowGamepadTextInput

(Of course, we don't link to the Steam API, so can't use these
directly.)

A future feature will be to provide a text input rect, so that keyboard
can position themselves out of the way, and other input method editors
can align any candidate lists correctly.