ck_play: Force screen refresh exiting ComputerWrist
This is present in the original DOS games, but wasn't in Omnispeak. I'd
assumed that the whole VHB_MarkUpdateBlocks() bit would take care of it,
but we're dirtying both buffers and then only marking one. So, alas, no.

This fixes a flickering black ring around the right and bottom edges
after returning from the menu on double-buffered backends.