id_in: Use all available input devices.
Add a new "all" control type which maps to all available input devices
(currently just Keyboard and first joystick), and force that to be used
in non-CK_VANILLA builds.

We already sort-of did this in the menu, but it's best to just do it
everywhere.

I may hide this behind a config option, but it's forced-default for now.