Initial SDL3 support
Add VL, SD, and IN backends for SDL3. We do this instead of using the
combined SDL1/2 backends for SD and IN, as so many functions have been
renamed. This does present some problems, as otherwise they share a lot
of code, so one will likely end up outdated.

Ideally, some of the 'generic' code, particularly in id_sd, can be split
out a bit.

One notable change is that all includes of "SDL.h" need to be replaced
with an include of "ck_cross.h", which has the correct logic to include
"SDL3/SDL.h" where needed. WITH_SDL is now set to 3 for SDL3 builds.

Note that, for Windows builds, SDL3 is not linked statically by default,
as SDL3 no longer provides mingw32 static builds:
https://github.com/libsdl-org/SDL/issues/11132

To use this, set RENDERER=sdl3.