ACTION.CK6: Use sprite #defines for Keen actions
Add all of the Keen sprites to EPISODE.CK6, and use them (as indirect
references) in ACTION.CK6 for all of the Keen actions.

Note that this handles both v1.4 and v1.5, though currently the ACTION.CK6
file is identical between the two (save for the version number). So there's
probably some save compat issues that've been around for ages.