ck_act: Add a debug feature to warn on use of unset variables
To make sure a variable isn't typoed, missing, etc, print a warning if an
unset variable is used, and therefore the default returned. This is enabled
with the CK_VAR_WARNONNOTSET #define, which is enabled by default in DEBUG
builds.

It's already caught lots of issues, so it's nice to have.