sd: Add infrastructure for detecting soundcards
We currently assume all backends support both PC Speaker and OPL2,
but different backends support different features in practice. On DOS,
maybe only the PC Speaker is available. The various hardware OPL backends
don't support PC Speaker (or emulate it badly), and at some point
may support OPL3 features (which we may want to use for future stereo
music support).

Add a "detect" callback which returns a bitmask of supported card features.
We don't do anything with it yet, but it is there.