ck_play: Add some missing RF_ForceRefresh() calls
Despite in theory it being possible for things like the Help Screen to properly
use VHB_ functions to not break normal gameplay rendering, they don't, so we
need to force a refresh when gameplay resumes. This is what the original game
code does.

Fixes some nasty issues where a ring of flashing red appeared after closing the
help screen.