id_vl_dos: A bunch of small optimisations for DOS/EGA
The DOS version of Omnispeak is much, much slower than Keen, especially
on older machines with slow EGA/VGA cards (e.g. ISA cards). There are
lots of reasons for this, but the VL_DOS code counts for most of them.

This change doesn't fix this (a real fix will probably require moving
the tile buffer to video RAM and having optimised tile blitters), but
does make a number of small improvements. It's still not really playable
on, e.g., the Hand386, but it's better than it was.

The improvements are:
- Yank a few calculations out of inner loops, and replace some
  multiplications with additions. (This was mostly already done by
  compiler optimisations, but it's bets to make sure.)
- Cache the EGA enabled plane and write mode to avoid slow port I/O.
- Improve masked blitting by ANDing the mask 4 planes at a time with the
  EGA ALU. This reduces the slow bus writes by ~37%.
- Add a (not yet used) optimised version of ScrollSurface which uses a
  single memmove(). This is slightly faster, but not actually copying
  would be better still. (As would moving this to video memory).