varparser: Add a tool to convert EPISODE.CKx variables to C
We may want to be able to compile variable values from EPISODE.CKx
(including strings, actions, chunk numbers, etc) into the final game
binary. This reduces memory use and parsing time, and would allow us to
more easily port mod code between the original DOS keen and Omnispeak.
However, it only works with one episode per build.

varparser is the first stage of a tool to do this. It uses CK_VAR to
parse the variables (in the debug mode with typechecking) and can output
a header full of #defines and a source file full of static data
(actions, potentially strings). These can then be compiled in.

An even hackier version of this was used for the DOS16 version of
Operation Station Infiltration, and mostly worked. This uses the
Atroxian Realm version of CK_action, which doesn't include the
compatDosOffset field, as well as generating functions to convert to and
from the expected DOS offsets.

Note we have to increase the size of the arena for the typechecking
info.