Initial support for compiling individual episodes
It's now possible to enable/disable support for Keens 4, 5, and 6 independently,
i.e., create a Keen4-only version of omnispeak or similar.

This can be useful when building versions for platforms without the command-line
options otherwise used to select an episode, or for mods (which don't need extraneous
code from other epissodes present. It will also be a prerequisite for compiling
variables in permanently (if we choose to support that in the future).

Use the WITH_KEEN{4,5,6} options in either make (make WITH_KEEN4=1 WITH_KEEN5=0) or
CMake (cmake -DWITH_KEEN4=ON -DWITH_KEEN5=OFF ..). All episodes default to on.