id_rf: Properly mark all tiles in non-looping animations
AT LONG LAST, THE BEAST IS DEAD!

Omnispeak has had a long-lasting bug where gem doors would cause an
error to be printed when opened, e.g.
"Warning: tried to render a Tile16m which was not cached (index = 1871, chunk = 3690)"

This is caused by RF_MarkTileGraphics failing to mark all of the tiles
used in the door opening animation for caching when the level loads.
Keygem doors (and some other things) are implemented as animated tiles,
with a couple of special properties:
1. The animation time for the first frame is 0, causing the animation to
   not trigger until the game code manually activates it (such as when
   the gem is placed).
2. The animation doesn't loop, i.e. the "next tile offset" for the final
   tile in the animation is "0": it animates to itself.

It is the second of these which was causing the issue.

When a level is loaded, RF_MarkTileGraphics searches for all animating
tiles in the level and goes through all of the tiles in the animation
loop, and marks all of them to be cached. It knows it's reached the end
of an animation loop when one of two things happens:
a) The next tile in the animation 'chain' is the same as the first tile
   (the one present in the level). This means that the animation has
   looped back to the beginning.
b) The next tile in the animation 'chain' is the same as the current
   tile in the animation chain (i.e., case '2' above). This is checked
   by verifying that the offset used to reach the next tile was not 0.

Our bug is with (b). Instead of checking that the next tile is the same
as the previous one (i.e., the offset used to reach it is zero), we
checked that the offset from the next tile to the one after it is zero,
meaning we end the loop one early and never cache the final tile in a
non-looping animation. Whoops!

We now store the offset we used to reach the current tile, and use that
for the check. This fixes the issue, and doors now open successfully. We
also fix the same bug with background animations, though I don't think
there are any non-looping ones in the original games.