ck_act: Add 'chunk id' macros, use for item shadows
Currently, we hardcode a number of chunk numbers throughout the game
code, using global variables and episode setup functions to handle
them differing between episodes. Some of these share a chunk name, and
can easily be replaced with a CK_CHUNKNUM() lookup, but others need
to be stored in an array or otherwise manipulated.

Add a new chunk_id_t type, which represents the "id" of a chunk. This
is either a chunk number or a string representing the chunk name.
We can keep an array of these chunk ids, using CK_CHUNKID() to create
one (will either be a string or number, depending on if chunks are
compiled in), and CK_LookupChunk() to convert a chunk_id_t to a chunk
number (either returning it straight, or looking up the variable).

To prove this works, we port the item shadows over to use it. Seems
like it works so far!