ck_map: Use chunk_id_t for mapkeen frames
The original game uses the first frame of each direction, minus 1.
However, that's really ugly to hardcode, so we just use the first frame,
and take 1 from all of the places it's used. I can't think of any reason
why this was done this way in the orignal game, but life is full of
surprises, isn't it?

This gets rid of an extra thing setup per-episode, which should be nice.