ck_keen: Allow disabling "pole door glitches"
Keen's yDirection needs to be Down (1) when on the ground, or we can hit
_nasty_ issues when going through a door, leading to a visual glitch
(walking down into the door) and an exploit (the door's target is read
from the wrong tile, becoming 0000 in most cases, leading to finishing
the level, and possibly going to the secret level in Keen 5).
See:
   https://keenwiki.shikadi.net/wiki/Keen_5_Bugs#Ampton_Pole_Push_Door_Trick
   https://keenwiki.shikadi.net/wiki/Keen_5_Bugs#Shelley.2FDoor_Trick
   https://keenwiki.shikadi.net/wiki/Keen_4_Bugs#Bounder_Pole_Door_Glitch

We permit players to disable this bug by setting ck_poleDoorGlitch =
false in OMNISPK.CFG. We do this by forcing obj->yDirection = 1 in
CK_KeenSpecialColFunc() when keen is pushed.