id_in: Fix UTF-8 hack with signed char
For text input, we're expecting a UTF-8 string, but only support ASCII.
However, we get the input as a const char *, but UTF-8 really needs
unsigned values. Since the signedness of char is implementation-defined,
we need to support both, but the previous code only really worked with
unsigned char. Whoops.

Clang complains, which is annoying:
id_in.c:261:18: warning: result of comparison of constant 128 with expression of type 'const char' is always true [-Wtautological-constant-out-of-range-compare]
        if (utf8Text[0] < 0x80)
            ~~~~~~~~~~~ ^ ~~~~
1 warning generated.

So hack around it.