us_2: Use chunk_id_t for menu headers.
This is one of the remaining sets of constants where we use the episode
DefineConstants function, an extra layer of pointer indirection, and a
boatload of globals to work around needing arrays of chunk ids.

Using chunk_id_t should mostly fix it here, though we may have some
trouble with the unused DEBUG menu header.