ck_def: Remove the global per-episode chunk numbers
We were previously storing a number of "chunk numbers" in global
variables, as they were #defines in the original game, and varied
per-episode, so we needed something we could change at runtime to
support multiple episodes in the same binary.

We now have chunk_id_t and CK_LookupChunk() to serve this purpose, and
have migrated all uses of these away, so we can delete all of these
globals.