ck_text: Don't compile code for disabled episodes
Keen 6 doesn't have a help screen. Only Keen 5 has the 'secret ending'.
However, even if no episode which needed them were compiled in, we
enabled both.

For the Keen 5 secret ending, just put it behind an #ifdef WITH_KEEN5.

For the helpscreen, add a new HAS_HELPSCREEN #define, and enable it if
we're building either Keen 4 or Keen 5. This will allow future Keen 6
mods to more easily add the screen back in. (Though there are a lot of
runtime checks, as well.)

This will become necessary if we want to link the variables, as
TEXT_SECRETEND, etc, won't be defined anywhere.