ck_act: Add integer array support
Add a new type to EPISODE.CKx, the integer array. This can be used to
encode a list of integers, which the game can either read as an
intptr_t*, or access an element at a time. If the latter, typechecking
can validate that the index is in bounds (enabled in debug builds).

This an be used for a number of things currently stored in arrays of
chunk_id_t, but also for things like the level music, and maybe the lump
system.