ck_var: Add "string arrays", and use for level names/entry text
The way we were previously handling string arrays was by concatenating
an number onto a variable name. This worked, (and in some ways did make
it easier to see which level a name was for), but was ultimately rather
ugly.

Since we have integer arrays, we can have the corresponding string
arrays for these. These are implemented in the same way, and should work
comfortably.