ck_obj: Move some platforms, turrets, etc behind WITH_ep defines
Some episodes of Keen, particularly Keen 4, are lacking a number of
platform types, turrets, etc. We shouldn't compile code for them in if
only Keen 4 is enabled.

This not only makes the executable smaller for single-episode builds,
but also will be required if we want to link actions in, as otherwise
we'd hit linker errors for actions which don't exist.