varparser: Add a hack to declare all functions used in actions
The --declare-funcs option will make headers include forward
declarations for any functions used in actions. This is very redundant
at the moment, as there's no de-duplication occurring, and it breaks
with CK_ObjBadstate, which is used both as a think/draw and collision
function, but should be enough to get per-episde builds working with
linked actions (with a few extra hacks).