tests: Use dumps from Keen 6 EGA v1.5, as that's what we have in CI
The Keen 6 version we're testing in CI is actually version 1.5, but the
dumps were taken from v1.4 (as that's what's on my old floppy disk).
This used to work, as the offsets were stored in the omnispeak episode
struct, which was independent of the data being used. But now it's all
in the data, running "v1.4" with data from "v1.5" won't work any more.

Add dumps from v1.5, and compare against them in CI, as that's what
we're actually testing. At some point, I'll add v1.4 to CI as well, so
we can verify both. :-)