id_rf: Honour at most one scroll block in each direction.
When scrolling up or to the left, we'd loop through all scroll blocks,
and _push_ the screen down or right by a tile to avoid it. However, if
there were two scroll blocks on the same row/column, we'd do this for
both of them, leading to twice the displacement. This caused the hill in
Dunesville in RTTS to make the camera shake, as keen would try to move
the camera upward, but it'd be pushed back twice as far.

This was due to some missing 'break' statements in
RF_SmoothScrollLimit(). Checked this against the RTTS code and
keensource456, as well as in-game, and it seems fine.