id_mm: Align arena allocations
MM_ArenaAlloc() only returns byte-aligned blocks of memory, but
technically some of the things we allocate need to be aligned elsewhere.

This isn't a problem on x86 (except if SSE is used, and it isn't), but
ubsan picks up on it.

Add MM_ArenaAllocAligned(), which also accepts a power-of-two alignment.
We then use that where ubsan complains.

There are a few things missing, though:
- The typechecked versions aren't aligned.
- There's no implementation for the debug arena. This is not a huge
  problem, as it uses malloc() which is guaranteed to return
  sufficiently aligned data.
- The alignment for most types is hardcoded to their sizeof(). We should
  use alignof(), but support older compilers, so will need a ck_cross()
  wrapper or similar.