ck_config: Add CK_NEW_FEATURE_DEFAULT
This macro is defined as true on non-vanilla builds, false otherwise.
It's used for setting the default for config options, where the
average user will want things (bugfixes which don't affect normal
gameplay, improved saving of settings, etc), but which are technically
not vanilla compatible.

Also use it for in_forceJoyMenu, which is the only thing so far to need it.