Add a bunch of debug log messages.
These not only give a "Doom startup log" style aesthetic to everything,
but also could be useful in debugging things. It also should look cool
if enabled on the 3DS builds, where the second screen has log messages.

They're not enabled for release builds by default, so shouldn't be a
problem, but can be disabled by settings "logLevel = warn" in the
OMNISPK.CFG file.

The weirdest message is that, when caching things, the loading string is
printed ("Keen enters\nthe Shadowlands", etc).

Note that FS_ and MM_ startup messages are printed _before_ the config
file is loaded, which can prevent them from being disabled/enabled with
the config option.