id_mm: Simplify MM_ArenaStrDup()
MSVC complains a lot about strcpy, and wants us to use strcpy_s instead:
warning C4996: 'strcpy': This function or variable may be unsafe. Consider using strcpy_s instead.

But, we actually already know the length of the buffer perfectly matches the length of the string (since we allocated it), so we can just use memcpy() and avoid both needing to check for '\0', and needing to add a new null terminator at the end.

And the warning goes away, too (though there are lots of others.)