ck_cross: Add a custom safer strcpy() variant
Keen has a lot of calls to strcpy() around the place, and several of them are into small, fixed-size buffers. While I'm not immediately aware of any which will overflow in normal, day-to-day gameplay, it's worrying to have them. And MSVC complains _a lot_ about any strcpy() call.

Introduce a new CK_Cross_strscpy() function, which acts mostly like the Linux kernel strscpy() function (i.e., it always null terminates, and never writes (nor reads!) more than the buffer size), except it will Quit() if the buffer is too small.

I may change this in the future to only Quit() in DEBUG builds, but it works for now.

Lots of uses of this to follow in the next patch!