ck_text: Fix MSVC warnings
MSVC complains that the Print Colour should be an int8_t, so fix that. (The index can be 'int', too. It shouldn't matter.)

Equally, get rid of all the manual text manipulation and use one snprintf(), to avoid potentially overflowing a buffer. (Not really possible with the fixed pg x of y and a 64-byte buffer, but better safe than sorry.)