id_us_2: ck_us_2: Fix a whole bunch of warnings
MSVC gave a huge number of warnings of these files. This commit fixes almost all of them.

It mostly consists of reworking string manipulation, but there's a few other bits an pieces.

The remaining warnings are for getenv() and localtime(). The use of getenv() in particular is somthing I should look into. The original game does this, buy why?