id_us_1: Fix MSVC warnings
There were a few of these, mostly string functions (replaces with 'safer' variants, or printfs), and potentially lossy type conversions (mostly fixed by using size_t instead of uint16_t in a few places).

The biggest change is with USL_XORICursor, which now accepts a size_t for the cursor position, not a uint16_t.