VL backend for the new SDL3 GPU API
This is a new VL render backend for the SDL3 GPU API.

It can be enabled with RENDERER=sdl3gpu, and only supports SPIR-V
shaders (so won't work with anything but the Vulkan API for now).

This also changes the SDL2 Vulkan backend to match the descriptor layout
required by SDL3, so we can reuse the shaders. Hopefully the changes
here (like avoiding samplers) will not adversely impact any Vulkan
drivers.