ck_map: Fix flickering on teleporter use with double-buffering
The teleporters in Keen 5 and 6 would flicker noticably when using a
double-buffered VL backend (like sdl2gl and dos).

This is because we were calling VL_Present() after RF_Refresh(), which
actually already calls VL_Present() for us. With double-buffering, this
results in us showing the old buffer, which had the previous location
(prior to teleporting) in it.

Get rid of the extra VL_Present() calls.

Signed-off-by: David Gow <david@ingeniumdigital.com>