ck6_obj1: Add a bunch of missing FixRefreshBuffers
When Keen interacts with the Grabbiter, as well as other world-map
interactions in Keen 6 which spawn dialog boxes, they would appear on
top of the _previous_ frame on double-buffered backends (dos, sdl2gl).
This could cause minor glitches, like enemies moving backwards or the
score box being misaligned.

Add the appropriate calls to VL_FixRefreshBuffers() (like the original
game), which copies the contents of the new frame over the old frame
before drawing the dialog.

Signed-off-by: David Gow <david@ingeniumdigital.com>