ck_obj: Implement turret shot persistence bug on CK6
Unlike the turrets in Keen 5 (and the darts in Keen 4), the turrets in
Keen 6 spawn shots which are OBJ_ACTIVE, and so persist even if
offscreen. (Keen 5's turret shots are OBJ_EXISTS_ONLY_ONSCREEN). This
can lead to the objarray getting full of offscreen shots, causing no
future shots (including those from Keen's neural stunner) to spawn.

This is not actually a great feature, but to maintain bug-for-bug
compatibility we retain it. It can be "fixed" in Keen 6 (or, if you're
so inclined, broken in Keen 5) by adjusting the ck_turretShotsPersist
variable in EPISODE.CK6, or in OMNISPK.CFG.

Thanks to NY00123 for bringing this to my attention, and K1n9_Duk3 for
noticing this in his decompilation.