ck_obj: Fix maximum door height
The maximum door height in Keen should be 50 tiles, not 48, accorting
to K1n9_Duk3's keensource456. Either way, it should be a documented #define,
not a hardcoded hex number, so that we can debug it better.

(Note that we check the door size against this maximum and quit if it's
too tall, but Keen just corrupts the stack.)