ck_quit: Add a printf-formatted Quit(), QuitF()
The Quit() function is used to terminate the game on a fatal error, and
accepts a single string as its message (the normal end screen is
displayed if the message is NULL). However, we often would like to
provide extra diagnostic information in these circumstances, such as the
filename we couldn't load, or similar. This is particularly important as
more data moves to being loaded or defined at runtime, so can't be baked
into static strings.

The new QuitF() function behaves exactly like Quit(), but the message is
now a printf-style format string, and additional arguments can be
passed in afterwards.

For example:
QuitF("Couldn't load file %s!", filename)