in: us_1: Don't write IN_ctrl_All to CONFIG.CKx (Fix #67)
The new IN_ctrl_All controller type, which signifies _all_ available
controllers should be read from, should never be written to CONFIG.CKx,
as CONFIG.CKx could be read by the original game, which doesn't
recognise it (and so disables all control — the opposite of what we
want).

Instead, keep track of the previous value in CONFIG.CKx, and write that
if the control type is set to IN_ctrl_All.

There's probably room for better tracking this in the UI, but it'll have
to wait a bit…

Thanks NY00123 for noticing this, and reporting it as #67.