ck_play: Add the MapCamera underwater level to EPISODE.CK4
The Well of Wishes level uses the world-map camera (which keeps Keen
centred all the time), instead of the in-level camera (which has a
larger buffer, and follows the ground Keen's walking on more). Since
Keen isn't walking on the ground, the simple one makes more sense here.

But mods may wish to change this level, so make it patchable in
EPISODE.CK4. Note that the world map is still hardcoded as 0, and this
only works if the episode is EP_CK4.