ck_act: Add support for "action aliases"
An "Action Alias" allows for an action variable to be set up as a
"pointer" or "alias" of another action. This allows us to separate out
places where particular actions are used programmatically so that they
can be modified in EPISODE.CKx or ACTION.CKx.

For example, if we wanted slugs to emit something other than slime, we
could imagine checking for "CK4_SlugPooAction" instead of
"CK4_ACT_SlugSlime0", and then setting an action alias, as so:
%action CK4_SlugPooAction = CK4_ACT_SlugSlime0

As long as CK4_ACT_SlugSlime0 had already been defined (i.e., this needs
to be at the bottom of ACTION or EPISODE), then CK4_SlugPooAction would
be usable wherever CK4_ACT_SlugSlime0 previously was. And it could
easily be changed.

The two actions are _the same_: only one copy is loaded into memory, and
they therefore share the same pointer. It's just an extra name for it in
the variable hashtable.