ck6: Make the creature question (copy protection) data driven
The Keen 6 "Creature Question" copy-protection mechanism should be:
a) Able to be enabled/disabled from EPISODE.CKx
b) Data-driven, so that mods can change the sprites/answers/etc

This patch lays the groundwork for (a) by letting it be disabled by
EPISODE.CKx, and moves much of the data to EPISODE.CKx, allowing (b).

It does not let non-Keen6 episodes _add_ copy protection yet, but that's
not particualrly desirable anyway.