ck_inter: Clear a byte to the right of Terminator plaques
The Terminator opening scrolls several "plaques" containing the credits
through, and always clears the free space under them. What wasn't clear
(and still might be wrong, as the assembly is not super-clear) is that
it also clears a byte to the right of the displayed plaque, in order to
counteract the scrolling.

This is not required for the original game's data, as there's always
enough spare blank space to the right of the bitmap to cover any
"droppings" left the previous frame, but it can cause problems either
with a very low framerate hitch (possible on multitasking OSes), or with
modded data. In particular, the popular "A MOD of an id Software
production" bitmap.