vl_dos: Improve ScrollSurface on EGA/DOS (see #55)
The DOS backend has many issues, both performance and correctness, which
need resolving. This change tackles two of them, both related to
ScrollSurface.

The first is a performance/smoothness issue: because we're always doing
the "SVGA compat" mode, we make periodic copies from one end of video
memory to the other. However, we always copied _both_ pages, but we only
really need to handle one. Update this to better place pages, and to
only copy one at a time. To ensure correctness, track the crtc offset
and update this after a copy. This _shouldn't_ be necessary in most
cases, but does get rid of potential overlap (which never actually
occurs) when a wrap occurs.

Secondly, fix an issue whereby the edge tiles from one page would
overwrite the currently visible opposite edge from the visible page.
This made a flickery, incorrect edge appear. This is due to there not
being a sufficient "gap" between the pages which we can use during
scolling. Keen handles this by having the pages actually be very large
(1024px wide). If we try this, though, we still get the problem, as VL
asssumes the entire page is valid and needs scrollling, not just the
actually used "port". If we hack around this by adding a gap, everything
works fine. But we then break the terminator intro, which needs more
memory as it's super-wide.

Hack around this with a "set gap" function on DOS. It's not great, and I
vaguely want to replace this by allowing a specific "gap" and "pitch" to
be specified on all surfaces" (so VL can be aware of separate used and
unused bits of memory), and maybe redo the interface enough to make the
tile buffer live in VRam and scroll efficiently (and copy efficiently).