ck_ep: Remove 'hasCreatureQuestion' from episode struct
Don't hardcode if an episode has a creature question, but rely entirely
on the value in EPISODE.CKx. This does make it easier to accidentally
skip the copy protection by using an old version of EPISODE.CKx, but
no-one cares, and it's better than the existing version, which gets you
a useless screen with placeholder strings.

In the process, it should (in theory) be possible to add a creature
question to another episode, but really, why would you?

The /NOCOPY option and friends now skip the creature question by telling
the game it's already been passed.

Also fix a misspelt variable, ck6_numCopyCreatures. Whoops!