ck6: Mark the KEYGEMS lump for gem-carrying Blooglets
Blooglets which carry a gem (i.e., blooglets which have an infoplane
value > 10) should mark both the blooglet and keygem lumps when spawned,
so that the relevant sprites are cached. This is important for levels
like the Bloog Control Center, where all the gems live in blooglets.

Not a showstopper — Omnispeak will cache them at the last minute if we
must — but definitely a bug.

Maybe one day we'll move more of the infoplane value / spawning stuff to
EPISODE.CKx — it's necessary for more complex mods anyway — but for now,
it's still hardcoded.