id_str: Properly detect a full table in STR_AddEntry()
While we detect a full table in debug mode by counting the number of
element, in theory we should detect this _anyway_ by returning false. We
didn't, as we didn't keep track of when we returned to the original hash
value. So we'd hang, rather than returning an error. Whoops!

Still, this works fine now, and we probably can (or should) get rid of
the DEBUG-only Quit() at some point, as the callers of STR_AddEntry()
actually tend to have better error handling here anyway.