id_str: Improve hashtable efficiency.
We add a lot of strings which end in consecutive numbers to our
hashtables, such as action names, which often have multiple actions
ending in a frame count, e.g.:
CK6_ACT_OrbatrixUncurl0, CK6_ACT_OrbatrixUncurl1,
CK6_ACT_OrbatrixUncurl2, etc.

The djb2 xor hash we were using would often give these consecutive
hashes, which caused long runs of consecutive entries in the hashtable
to be used. This made collisions much worse, as there was a strong
likelihood that we'd end up with long chains of used entries before a
gap which could stop a search.

Multiply a simple bit-mixing constant to shake things up a bit. This
seems to drastically reduce the number of reported collisions.

In the process, make two other fixes to the hash:
- Start it with the value 5381, rather than zero, as in theory we
  should.
- Don't call strlen() separately, to avoid a second loop. (This is
  actually slower on some modern machines, but should be faster on very
  old machines, which is where we care about performance most.)