vl_dos: Fix build when BUILDASCPP=0
It turns out that ternary operators aren't lvalues in C, but are in
C++[1].
Since BUILDASCPP=1 by default for the makefile, I didn't notice that I'd
broken VL_DOS_ScrollSurface()'s build.

Fix it with an ugly if(). Obviously, we should just have data be an
array, not a couple of separate pointers, but meh. This'll all be
cleaned up one day anyway.

[1]: https://www.yhi.moe/blog/en/can-ternary-operator-be-an-lvalue-in-c-or-not