ck_ep: Load _all_ action functions for enabled episodes
Instead of only loading functions for the selected episode, load them
for _all_ enabled episodes. This allows ACTION.CKx to use functions from
any episode, at the cost of increasing the size of the function tables.

As a result, we can remove one more thing from the episode struct.

There are some downsides to this, so I may make future versions of
EPISODE.CKx able to request (or disable) these, either explicitly, or
depending on the base episode.

We also drastically increase the size of the variable hashtable, as we
were getting a collision rate of ~50% with 2048. With 8192, it's < 10%
(4096 ~= 20%). This is reasonably wasteful (64k), but 32-bit systems
_should_ have that to spare.