ck_main: ck_ep: Make /EPISODE take a filename
This is a deceptively major change, after which /EPISODE should take the
filename of the core EPISODE.CKx file, rather than just a number. This
allows mods to have their own names (and potentially just %include the
original game's data), as well as different variants (K6V14.CK6,
K6V15.CK6 anyone?).

The secret behind this is the new variable: ck_episodeNumber. This is
used to select the episode struct at the moment, but may end up
replacing it entirely down the line.

For compatibility:
- If an episode number (4,5,6,6e14,6e15) is passed to /EPISODE, a
  hardcoded EPISODE.CKx will be loaded.
- If the file provided doesn't have a ck_episodeNumber variable present,
  Omnispeak will attempt to guess from the file extension. But please,
  put ck_episodeNumber in.