Abiathar '.aslev' support
This is a quick implementation of loading levels from Abiathar .aslev
files. The bulk of it is an implementation of the Keen1 LZW algorithm.
This implementation is a bit different from most of the existing ones
(the 'dictionary' stores references to the original decompressed data),
but it should work pretty well — there's definitely the possibility of
growing the dictionary dynamically, though, which might be necessary to
support the absolutely huge potential dictionary sizes in .aslev files
(24 bit codes!)

It also loads .aslev files as 'Omni' files, and so they must be located
int the "omni" directory (alongside the game header files, like EGAHEAD,
etc.)

To use this, run Omnispeak with the /ASLEV [filename] option. This is
analagous to the /TEDLEVEL option, but loads the level (as level 1) from
the provided .aslev file.

The two issues (already mentioned above) are:
- .aslev files must live in the 'omni' directory, al;ongside EGAHEAD, etc.
- Loading .aslev files requires a lot of memory (256MB) to store the LZW
  dictionary, and so likely will not work on lower-end systems (e.g. DOS).

Both of those issues may be fixed in future versions... we'll see.

In the meantime, enjoy!