[WIP] Abiathar '.aslev' support
This is a quick implementation of loading levels from Abiathar .aslev
files. The bulk of it is an implementation of the Keen1 LZW algorithm.
This implementation is a bit different from most of the existing ones
(the 'dictionary' stores references to the original decompressed data),
but it should work pretty well — there's definitely the possibility of
growing the dictionary dynamically, though, which might be necessary to
support the absolutely huge potential dictionary sizes in .aslev files
(24 bit codes!)

Most of the actual bugs are due to needing to reload the file, which we
don't do, and needing a level number (hardcoded to 1). So this
definitely isn't stable. But it can be fun to play with, particularly
since none of the level data is reloaded on death, so all the platforms
disappear, anything which was collected or changed (e.g. switches) stays
collected, and there are other strange bugs.

Enjoy!