id_ca: Support uncompressed audio at runtime
We previously supported uncompressed audio via a compile-time #define,
CA_AUDIOUNCOMPRESSED, but this was annoying for two reasons:
- It couldn't be used without a tediout manual source change and
  recompile.
- It only supported the MUSE AUDIOT format, not the more common Keen Mod
  use of NULL huffman dictionaries, which still store the "expanded
  length" in the chunk.

This change adds a new "ca_audioCompressed" integer option, which can
take one of three values:

0: Huffman compressed audio, à la the original games
1: Uncompressed AUDIOT audio, as used by Bio Menace, Wolf3D, and some
   source mods.
2: NULL-Huffman-style uncompressed audio, which is used by a number of
   Keen mods.

Just throw this in EPISODE.CKx and have fun. If audio is not compressed,
there's no need to have a dictionary lying about.

Signed-off-by: David Gow <david@ingeniumdigital.com>